apps/ai-agent/requirements.txt
streamlit>=1.40.0
crewai>=0.80.0
crewai-tools>=0.14.0
python-dotenv>=1.0.0
